[YS-413] [YS-409] refactor: 공용으로 사용되는 ButtonInput 디렉토리 위치 변경
