+++ src/app/join/components/Participant/JoinEmailStep/JoinEmailStep.tsx
@@ -10,8 +10,8 @@ import JoinInput from '../../JoinInput/JoinInput';
 import useCheckValidEmailInfoQuery from '@/app/join/hooks/useCheckValidEmailInfoQuery';
 import useServiceAgreeCheck from '@/app/join/hooks/useServiceAgreeCheck';
 import { joinContentContainer, joinForm, nextButton } from '@/app/join/JoinPage.css';
-import ButtonInput from '@/app/user/profile/components/ButtonInput/ButtonInput';
 import { ParticipantJoinSchemaType } from '@/schema/join/ParticipantJoinSchema';
+import ButtonInput from '@/components/ButtonInput/ButtonInput';
 
 interface JoinEmailStepProps {
   onNext: () => void;
